# [risk=no][no ticket] Fix RTL performance flake