Attempts to fix the RTL performance flake described [in Slack](https://pmi-engteam.slack.com/archives/C7JUTPQLE/p1693936544686529?thread_ts=1693936143.040119&cid=C7JUTPQLE).

I had run into performance issues using `type` when originally converting these tests, and needed to change some instances to a `click` / `paste` interaction, instead. This PR changes all instances to that, since none of the tests specifically depend on "typing."

This decreases the runtime of the two tests in question from 1.7 seconds to 0.9 seconds on my machine. As a bonus, I updated the rest of the file, decreasing the total file runtime from 12.5 seconds to 6.8 seconds.

I wasn't able to reproduce the flake locally, since those tests ran in under 5 seconds on my machine. If this issue persists I'll dig deeper.

---
**PR checklist**

- [x] I have included an issue ID or "no ticket" in the PR title as outlined in [CONTRIBUTING.md](https://github.com/all-of-us/workbench/blob/main/.github/CONTRIBUTING.md).
- [x] I have included a risk tag of the form `[risk=no|low|moderate|severe]` in the PR title as outlined in [CONTRIBUTING.md](https://github.com/all-of-us/workbench/blob/main/.github/CONTRIBUTING.md).
- [x] I have manually tested this change and my testing process is described above.
- [x] This change includes appropriate automated tests, and I have documented any behavior that cannot be tested with code.
- [x] I have added explanatory comments where the logic is not obvious.
- One or more of the following is true:
  - [ ] This change is intended to complete a JIRA story, so I have checked that all AC are met for that story.
  - [x] This change fixes a bug, so I have ensured the steps to reproduce are in the Jira ticket or provided above.
  - [ ] This change impacts deployment safety (e.g. removing/altering APIs which are in use), so I have documented the impacts in the description.
  - [ ] This change includes a new feature flag, so I have created and linked new JIRA tickets to (a) turn on the feature flag and (b) remove it later.
  - [ ] This change modifies the UI, so I have taken screenshots or recordings of the new behavior and notified the PO and UX designer in [Slack](https://pmi-engteam.slack.com/archives/C02MWP2RN5P).
  - [ ] This change modifies API behavior, so I have run the relevant E2E tests locally because API changes are not covered by our PR checks.
  - [ ] None of the above apply to this change.